pyproject.toml
[tool.poetry]
name = "falkordb_gemini_kg"
version = "0.1.2-beta"
license = "MIT"
description = ""
authors = ["FalkorDB <info@falkordb.com>"]
readme = "README.md"
repository = "https://github.com/FalkorDB/falkordb-gemini-kg"
homepage = "http://RAG-SDK.readthedocs.io"
keywords = ['RAG', 'Graph-RAG', 'KnowledgeGraph']

[tool.poetry.dependencies]
python = "^3.11.4"
falkordb = "^1.0.4"
typing-extensions = "^4.12.1"
bs4 = "^0.0.2"
sqlalchemy = "^2.0.30"
pypdf = "^4.2.0"
vertexai = "^1.49.0"
backoff = "^2.2.1"
python-abc = "^0.2.0"
ratelimit = "^2.2.1"
python-dotenv = "^1.0.1"
openai = "^1.35.9"
fix-busted-json = "^0.0.18"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.1"

[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"
sphinx-rtd-theme = "^2.0.0"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
